+++ CHANGELOG.md
@@ -43,6 +43,8 @@ BUG FIXES:
   * providers/aws: Ingress results are treated as a set, so order doesn't
       matter anymore. [GH-87]
   * providers/aws: Instance security groups treated as a set [GH-194]
+  * providers/aws: Retry Route53 requests if operation failed because another
+      operation is in progress [GH-183]
   * providers/heroku: If you delete the `config_vars` block, config vars
       are properly nuked.
   * providers/heroku: Domains and drains are deleted before the app.